added some more test cases to test the functionality of downloader and download_thread

+++ tests/test_downloader.py
@@ -137,7 +137,8 @@ class BasicsTestCase(unittest.TestCase):
         test_sha = "QmREyeWxAGtuQ5UiiTs13zp5ZamjkVBYpnDCF1bTgn7Atc"
         print("Testing the create_html_from_url method to verify IPFS gets the file and it is renamed to have a .html "
               "ending.")
-        os.remove(down.basePath + test_sha + ".html")
+        if os.path.exists(down.basePath + test_sha + ".html"):
+            os.remove(down.basePath + test_sha + ".html")
         print("    There is a file called " + test_sha + ".html (should be False): " + str(os.path.exists(
             down.basePath + test_sha + '.html')))
         # :param test_sha: this is the IPFS hash of the example.html content
@@ -148,3 +149,9 @@ class BasicsTestCase(unittest.TestCase):
         print("    There is a file called " + test_sha + ".html: " + str(os.path.exists(
             down.basePath + test_sha + '.html')))
         self.assertTrue(os.path.exists(down.basePath + test_sha + '.html'))
+
+    def test_distributed_timestamp_not_none(self):
+        down.basePath = '/home/sebastian/testing-stw/'
+        result = down.distributed_timestamp("http://www.sueddeutsche.de/wirtschaft/oelpreis-saudischer-oelminister-die"
+                                            "-oelflut-ist-zu-ende-1.3047480")
+        self.assertIsNotNone(result)
